Added login implementation
You can now log in and the app with autheticate or create a new account
in firebase. App will then move to the main menu.

I’ve also created a view controller for the main menu

There’s a couple of edge cases and error handling I still need to test
out, but right now it’s a functioning feature.