Location Tracking is now enabled
When the user enters the bubble, the app grabs the events the user has
currently joined for the day and starts a timer. When the user walks
out, the app shows a notification that asks which event he/she did, and
updates the value in firebase (in seconds).

TODO:

When deleting an event in calendar, also delete it in participants
firebase

When finished with an activity in location tracking, make it so you
can’t update the values twice?